+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - My missions page allows user to search missions they contributed to, in addition to the ones they created
 
+### Changed
+- Move feature button on map of review page is more visible
+
 
 ## 0.5.7 - 2019-04-27